+++ src/EventListener/MarkdownResponseListener.php
@@ -3,13 +3,13 @@
 namespace TomvdPeet\MarkdownNegotiationBundle\EventListener;
 
 use League\HTMLToMarkdown\HtmlConverterInterface;
-use Symfony\Component\HttpFoundation\AcceptHeader;
 use Symfony\Component\HttpFoundation\BinaryFileResponse;
-use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\StreamedResponse;
 use Symfony\Component\HttpKernel\Event\ResponseEvent;
 use TomvdPeet\MarkdownNegotiationBundle\Html\HtmlCleaner;
+use TomvdPeet\MarkdownNegotiationBundle\Http\MarkdownNegotiator;
+use TomvdPeet\MarkdownNegotiationBundle\Http\MarkdownResponse;
 use TomvdPeet\MarkdownNegotiationBundle\Routing\MarkdownRouteMap;
 
 class MarkdownResponseListener
@@ -18,6 +18,7 @@ class MarkdownResponseListener
         private readonly MarkdownRouteMap $markdownRouteMap,
         private readonly HtmlConverterInterface $converter,
         private readonly HtmlCleaner $htmlCleaner,
+        private readonly MarkdownNegotiator $markdownNegotiator,
     ) {
     }
 
@@ -42,7 +43,7 @@ class MarkdownResponseListener
             return;
         }
 
-        if (!$this->prefersMarkdown($request)) {
+        if (!$this->markdownNegotiator->prefersMarkdown($request)) {
             return;
         }
 
@@ -56,7 +57,7 @@ class MarkdownResponseListener
         }
 
         $response->setContent($this->converter->convert($this->htmlCleaner->clean($content, $request->getUri())));
-        $response->headers->set('Content-Type', 'text/markdown; charset='.($response->getCharset() ?: 'UTF-8'));
+        $response->headers->set('Content-Type', MarkdownResponse::CONTENT_TYPE.'; charset='.($response->getCharset() ?: 'UTF-8'));
         $this->addVaryAccept($response);
     }
 
@@ -75,21 +76,7 @@ class MarkdownResponseListener
 
     private function isMarkdownResponse(Response $response): bool
     {
-        return 'text/markdown' === $this->getContentType($response);
-    }
-
-    private function prefersMarkdown(Request $request): bool
-    {
-        $accept = AcceptHeader::fromString($request->headers->get('Accept'));
-        $markdown = $accept->get('text/markdown');
-
-        if (null === $markdown || $markdown->getQuality() <= 0.0) {
-            return false;
-        }
-
-        $html = $accept->get('text/html');
-
-        return null === $html || $markdown->getQuality() > $html->getQuality();
+        return MarkdownResponse::CONTENT_TYPE === $this->getContentType($response);
     }
 
     private function getContentType(Response $response): ?string